* CU-2atcgy0_Upload-images_Osei-Quashie:
  CU-2atcgy0 - Creat migration and model for attachments - Get fill uploader and make into livewire component - Work on attachment functionality in the backend - Add attachment component to new post box

Conflicts:
	Modules/Social/Database/Migrations/2022_02_17_221205_create_posts_table.php
	Modules/Social/Http/Livewire/NewPostBox.php
	Modules/Social/Resources/views/livewire/new-post-box.blade.php

+++ Modules/Social/Database/Migrations/2022_02_17_221205_create_posts_table.php
@@ -19,7 +19,7 @@ class CreatePostsTable extends Migration
             $table->unsignedBigInteger('team_id')->nullable();
             $table->text('body');
             $table->text('type')->nullable();
-            $table->integer('postable_id')->nullable();
+            $table->unsignedBigInteger('postable_id')->nullable();
             $table->string('postable_type')->nullable();
             $table->timestamps();
         });
